Move GIL and some threading code out of ceval.c into gil.c.
