+++ Tools/gdb/libpython.py
@@ -1752,7 +1752,7 @@ class Frame(object):
 
     def is_waiting_for_gil(self):
         '''Is this frame waiting on the GIL?'''
-        # This assumes the _POSIX_THREADS version of Python/ceval_gil.h:
+        # This assumes the _POSIX_THREADS version of Python/gil.c:
         name = self._gdbframe.name()
         if name:
             return (name == 'take_gil')
